filter_contours_area_of_image*: also ensure validity here

+++ src/eynollah/eynollah.py
@@ -3671,13 +3671,13 @@ class Eynollah:
 
     def dilate_textregions_contours_textline_version(self, all_found_textline_polygons):
         return [[np.array(make_valid(Polygon(poly[:, 0]).buffer(5)).exterior.coords[:-1],
-                          dtype=int)[:, np.newaxis]
+                          dtype=np.uint)[:, np.newaxis]
                  for poly in region]
                 for region in all_found_textline_polygons]
 
     def dilate_textregions_contours(self, all_found_textline_polygons):
         return [np.array(make_valid(Polygon(poly[:, 0]).buffer(5)).exterior.coords[:-1],
-                         dtype=int)[:, np.newaxis]
+                         dtype=np.uint)[:, np.newaxis]
                 for poly in all_found_textline_polygons]
